refactor(#93): sidebar을 덮는 형식으로 수정
