+++ src/components/ChartBox/KeywordChartBox.jsx
@@ -3,8 +3,8 @@ import { Tabs, Tab } from 'react-bootstrap';
 import CandleChart from './CandleChart';
 import CandleChartSimple from '../CandleChartSimple';
 import ChartData from './ChartData';
-import UserLike from './UserLike';
 import UserKeywordLike from './UserKeywordLike';
+import { useOutletContext } from 'react-router-dom';
 
 export default function KeywordChartBox({
   chartData,
@@ -19,6 +19,7 @@ export default function KeywordChartBox({
   const chartContainerRef = useRef(null); // CandleChart 상위 div 참조
   const [containerDimensions, setContainerDimensions] = useState({ width: 0, height: 0 });
   const [chartDimensions, setChartDimensions] = useState({ width: 0, height: 0 });
+  const { isSidebarOpen } = useOutletContext(); // isSidebarOpen 상태 가져오기
 
   useEffect(() => {
     bringStockChart(stockInfo.stock_code, setChartData, period);
@@ -42,7 +43,7 @@ export default function KeywordChartBox({
     handleResize();
     window.addEventListener('resize', handleResize);
     return () => window.removeEventListener('resize', handleResize);
-  }, []);
+  }, [isSidebarOpen]);
 
   // CandleChart 상위 div 크기 측정
   useEffect(() => {
@@ -58,7 +59,7 @@ export default function KeywordChartBox({
 
     if (chartContainerRef.current) observer.observe(chartContainerRef.current);
     return () => observer.disconnect();
-  }, []);
+  }, [isSidebarOpen]);
 
   const currData = chartData[chartData.length - 1];
